+++ packages/swirl-components/src/index.ts
@@ -1,5 +1,9 @@
 export { SwirlThemeChangeEventData } from "./components/swirl-theme-provider/swirl-theme-provider";
 export { SwirlAppLayoutMobileView } from "./components/swirl-app-layout/swirl-app-layout";
+export {
+  SwirlAutocompleteValue,
+  SwirlAutocompleteSuggestion,
+} from "./components/swirl-autocomplete/swirl-autocomplete";
 export { WcDatepickerCustomEvent } from "wc-datepicker";
 export {
   WCDatepickerLabels,