+++ apps/swirl-docs/CHANGELOG.md
@@ -1,5 +1,61 @@
 # swirl-docs
 
+## 0.1.170
+
+### Patch Changes
+
+- Updated dependencies
+  [[`f8ed5afd`](https://github.com/getflip/swirl/commit/f8ed5afd1eff954bbc3fc68ae9bd3c9f00b58bfc)]:
+  - @getflip/swirl-components@0.87.5
+  - @getflip/swirl-components-react@0.86.5
+
+## 0.1.169
+
+### Patch Changes
+
+- Updated dependencies
+  [[`83bb0b97`](https://github.com/getflip/swirl/commit/83bb0b9748ea84ed7b6635da936bfe75d5f161e5)]:
+  - @getflip/swirl-components@0.87.4
+  - @getflip/swirl-components-react@0.86.4
+
+## 0.1.168
+
+### Patch Changes
+
+- Updated dependencies
+  [[`1c6cb1d9`](https://github.com/getflip/swirl/commit/1c6cb1d91daf726142f1d445a7a3e755d01cd333)]:
+  - @getflip/swirl-components@0.87.3
+  - @getflip/swirl-components-react@0.86.3
+
+## 0.1.167
+
+### Patch Changes
+
+- Updated dependencies
+  [[`c6032d45`](https://github.com/getflip/swirl/commit/c6032d45cda2e0ff7fcedf0537095e2a2fdd207b)]:
+  - @getflip/swirl-components@0.87.2
+  - @getflip/swirl-components-react@0.86.2
+
+## 0.1.166
+
+### Patch Changes
+
+- Updated dependencies
+  [[`0356bfb5`](https://github.com/getflip/swirl/commit/0356bfb5211483880a0f2b83e0a0b5a6e068fd67)]:
+  - @getflip/swirl-components@0.87.1
+  - @getflip/swirl-components-react@0.86.1
+
+## 0.1.165
+
+### Patch Changes
+
+- Updated dependencies
+  [[`67fa3772`](https://github.com/getflip/swirl/commit/67fa3772909b63701dbd0801feb15fe07eaacb85),
+  [`67fa3772`](https://github.com/getflip/swirl/commit/67fa3772909b63701dbd0801feb15fe07eaacb85),
+  [`67fa3772`](https://github.com/getflip/swirl/commit/67fa3772909b63701dbd0801feb15fe07eaacb85)]:
+  - @getflip/swirl-components@0.87.0
+  - @getflip/swirl-components-react@0.86.0
+
 ## 0.1.164
 
 ### Patch Changes