+++ packages/swirl-components-react/CHANGELOG.md
@@ -1,5 +1,92 @@
 # @getflip/swirl-components-react
 
+## 0.86.5
+
+### Patch Changes
+
+- [`f8ed5afd`](https://github.com/getflip/swirl/commit/f8ed5afd1eff954bbc3fc68ae9bd3c9f00b58bfc)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Let swirl-text-input use full
+  width when used without a suffix
+
+- Updated dependencies
+  [[`f8ed5afd`](https://github.com/getflip/swirl/commit/f8ed5afd1eff954bbc3fc68ae9bd3c9f00b58bfc)]:
+  - @getflip/swirl-components@0.87.5
+
+## 0.86.4
+
+### Patch Changes
+
+- [`83bb0b97`](https://github.com/getflip/swirl/commit/83bb0b9748ea84ed7b6635da936bfe75d5f161e5)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Fix an issue closing the
+  swirl-autocomplete suggestions
+
+- Updated dependencies
+  [[`83bb0b97`](https://github.com/getflip/swirl/commit/83bb0b9748ea84ed7b6635da936bfe75d5f161e5)]:
+  - @getflip/swirl-components@0.87.4
+
+## 0.86.3
+
+### Patch Changes
+
+- [`1c6cb1d9`](https://github.com/getflip/swirl/commit/1c6cb1d91daf726142f1d445a7a3e755d01cd333)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Fix swil-option-list focus and
+  drag and drop issues
+
+- Updated dependencies
+  [[`1c6cb1d9`](https://github.com/getflip/swirl/commit/1c6cb1d91daf726142f1d445a7a3e755d01cd333)]:
+  - @getflip/swirl-components@0.87.3
+
+## 0.86.2
+
+### Patch Changes
+
+- [`c6032d45`](https://github.com/getflip/swirl/commit/c6032d45cda2e0ff7fcedf0537095e2a2fdd207b)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Fix swirl-button "form" prop
+
+- Updated dependencies
+  [[`c6032d45`](https://github.com/getflip/swirl/commit/c6032d45cda2e0ff7fcedf0537095e2a2fdd207b)]:
+  - @getflip/swirl-components@0.87.2
+
+## 0.86.1
+
+### Patch Changes
+
+- [`0356bfb5`](https://github.com/getflip/swirl/commit/0356bfb5211483880a0f2b83e0a0b5a6e068fd67)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Limit swirl-autocomplete listbox
+  height
+
+- Updated dependencies
+  [[`0356bfb5`](https://github.com/getflip/swirl/commit/0356bfb5211483880a0f2b83e0a0b5a6e068fd67)]:
+  - @getflip/swirl-components@0.87.1
+
+## 0.86.0
+
+### Minor Changes
+
+- [#359](https://github.com/getflip/swirl/pull/359)
+  [`67fa3772`](https://github.com/getflip/swirl/commit/67fa3772909b63701dbd0801feb15fe07eaacb85)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add multi select support to
+  swirl-autocomplete
+
+* [#359](https://github.com/getflip/swirl/pull/359)
+  [`67fa3772`](https://github.com/getflip/swirl/commit/67fa3772909b63701dbd0801feb15fe07eaacb85)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Breaking: Change
+  swirl-autocomplete value type to SwirlAutocompleteSuggestion or
+  SwirlAutocompleteSuggestion[]
+
+### Patch Changes
+
+- [#359](https://github.com/getflip/swirl/pull/359)
+  [`67fa3772`](https://github.com/getflip/swirl/commit/67fa3772909b63701dbd0801feb15fe07eaacb85)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Fix swirl-select keyboard
+  controls
+
+- Updated dependencies
+  [[`67fa3772`](https://github.com/getflip/swirl/commit/67fa3772909b63701dbd0801feb15fe07eaacb85),
+  [`67fa3772`](https://github.com/getflip/swirl/commit/67fa3772909b63701dbd0801feb15fe07eaacb85),
+  [`67fa3772`](https://github.com/getflip/swirl/commit/67fa3772909b63701dbd0801feb15fe07eaacb85)]:
+  - @getflip/swirl-components@0.87.0
+
 ## 0.85.0
 
 ### Minor Changes